requirements.txt
imdbpy
tqdm
joblib
pandas
numpy
altair
seaborn
streamlit
datetime
ipywidgets
pytrends
plotly
bokeh
wordcloud